Silently modify projects for lein javac
Previously, in order to modify projects so that they would compile
correctly with `lein javac`, a new profile was created and merged in.
When `project.clj` defined `:target-path` with a "%s", the formatted
build directory would then reference the additional profile that was
merged in. This causes ClassNotFoundExceptions in extreme cases, so now
the profile is simply modified directly before being used.